refactor: move LoginForm to components folder
