+++ src/pages/Login/Login.tsx
@@ -1,4 +1,4 @@
-import { LoginForm } from './LoginForm';
+import { LoginForm } from './components/LoginForm';
 import { Header } from '@/components';
 
 export const Login = () => {
