+++ src/pages/Login/index.ts
@@ -1,1 +1,1 @@
-export { Login } from './Login';
+export * from './Login';
